server/services/netsuite.ts
import { db } from "@db";
import { apiCredentials, apiStatus } from "@db/schema";
import { eq } from "drizzle-orm";

export class NetSuiteService {
  private static instance: NetSuiteService;
  private credentials: any;

  private constructor() {}

  static async getInstance() {
    if (!NetSuiteService.instance) {
      NetSuiteService.instance = new NetSuiteService();
      await NetSuiteService.instance.initialize();
    }
    return NetSuiteService.instance;
  }

  private async initialize() {
    const creds = await db.query.apiCredentials.findFirst({
      where: eq(apiCredentials.service, "netsuite")
    });
    
    if (!creds) {
      throw new Error("NetSuite credentials not found");
    }
    
    this.credentials = creds.credentials;
  }

  async updateStatus(status: "healthy" | "degraded" | "down", error?: string) {
    await db.insert(apiStatus).values({
      service: "netsuite",
      status,
      errorMessage: error,
      lastCheck: new Date()
    }).onConflictDoUpdate({
      target: [apiStatus.service],
      set: {
        status,
        errorMessage: error,
        lastCheck: new Date()
      }
    });
  }

  async findCustomerByPhone(phoneNumber: string) {
    try {
      // Make API call to NetSuite to find customer
      // Return customer data if found
      return {
        name: "John Doe",
        email: "john@example.com",
        purchaseHistory: []
      };
    } catch (error) {
      console.error("Error finding customer:", error);
      throw error;
    }
  }

  async getPurchaseHistory(customerId: string) {
    try {
      // Make API call to NetSuite to get purchase history
      return [];
    } catch (error) {
      console.error("Error getting purchase history:", error);
      throw error;
    }
  }
}
